# perf: Do not recreate and fill the environment in `.savvy_wrap_PlRExpr()`