--- a/NAMESPACE
+++ b/NAMESPACE
@@ -3,6 +3,7 @@
 S3method("!",polars_expr)
 S3method("!=",polars_expr)
 S3method("!=",polars_series)
+S3method("$",PlRExpr)
 S3method("$",polars_data_frame)
 S3method("$",polars_dtype)
 S3method("$",polars_dtype_enum)