Fixes #1439 

Just a PoC for now, would need to automate it to automatically grab the content of `.savvy_wrap_PlRExpr()` in 000-wrappers.R and put it in 001-wrappers.R

@eitsupi feel free to supersede this PR if you want, you're more familiar with this technique since you did https://github.com/eitsupi/neo-r-polars/pull/265/

With this PR:

```r
bench::mark(
  old = polars0::pl$col("a"),
  new = pl$col("a"),
  check = FALSE,
  iterations = 1e4
)

#> # A tibble: 2 × 13
#>   expression    min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr> <bch:> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 old        70.2µs 78.4µs    11865.        0B     14.3  9988    12      842ms
#> 2 new        36.9µs   39µs    23857.      560B     19.1  9992     8      419ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>
```