@eitsupi 2025-07-21
Trying this:

```r
bench::mark(
  call_cal = pl$col("a"),
  check = FALSE,
  iterations = 1e4
)
```

on current main (a9102a92a00a43c68f2e0cdb52397c947c929646)

```
# A tibble: 1 × 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory     time      
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list>     <list>    
1 call_cal     72.3µs   84.7µs    10699.    14.1KB     17.1  9984    16      933ms <NULL> <Rprofmem> <bench_tm>
# ℹ 1 more variable: gc <list>
```

on this PR (d6df3d0c97be8c051babb80b931f3ee302462fa5)

```
# A tibble: 1 × 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory              time       gc      
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list>              <list>     <list>  
1 call_cal      486µs    610µs     1507.    15.3KB     9.55  9937    63      6.59s <NULL> <Rprofmem [15 × 3]> <bench_tm> <tibble>
```

(Both have the same Rust lib and is debug build.)