@yutannihilation 2025-07-21
Curious! Btw, I'm wondering if I should switch from the current hand-crafted pseudo class to S7 at some point. Do you see any performance concerns?