@eitsupi 2025-07-21
> I'm wondering if I should switch from the current hand-crafted pseudo class to S7 at some point. Do you see any performance concerns?

Interesting. I don't know about performance except this article <https://rconsortium.github.io/S7/articles/performance.html>.

I am disappointed that the S7 currently does not have the ability to register associated functions as properties (RConsortium/S7#522).