@etiennebacher 2025-07-22
Thanks @eitsupi, I've tried with my test dataset in tidypolars and this reduces a lot of the slowdown observed with polars 1.0.0.