@etiennebacher 2025-07-22
@eitsupi do you want to keep copilot automatic review? It mostly adds some noise to me so I wouldn't mind deactivating it.