@eitsupi 2025-07-22
> Could it be included in a patch release 1.0.1 with `[[.polars_data_frame`?

To avoid the troublesome lib release, I cherry-picked only the changes on the R side and opened #1448.

> do you want to keep copilot automatic review?

I would like to leave it like this because it often provides useful feedback straight away.